Replace default decoder.
This commit changes the defacto default decoder to be the succinct based
decoder, with the traversal based decoder moved to a separate module.

Documentation hasn't made it over yet.

Updated bounds on bytestring.
