+++ test/Main.hs
@@ -6,7 +6,7 @@ import qualified Control.Lens                as L
 
 import           Data.Either                 (isLeft)
 
-import qualified Data.Scientific as Sci
+import qualified Data.Scientific             as Sci
 
 import           Data.ByteString             (ByteString)
 import qualified Data.ByteString.Builder     as BB
@@ -26,11 +26,16 @@ import           Test.Tasty
 import           Test.Tasty.Hedgehog
 import           Test.Tasty.HUnit
 
+import Data.Digit (HeXDigit)
+
 import           Waargonaut                  (Json)
 import qualified Waargonaut                  as W
 import qualified Waargonaut.Types.CommaSep   as CommaSep
+
+import Waargonaut.Types.JChar      (JChar)
 import qualified Waargonaut.Types.JChar      as JChar
-import qualified Waargonaut.Types.JNumber      as JNumber
+
+import qualified Waargonaut.Types.JNumber    as JNumber
 import qualified Waargonaut.Types.Whitespace as WS
 
 import qualified Waargonaut.Decode           as D
@@ -90,14 +95,21 @@ prop_uncons_consCommaSepVal = property $ do
 prop_jchar :: Property
 prop_jchar = property $ do
   c <- forAll Gen.unicodeAll
-  Just c === ((JChar._JChar #) <$> (c ^? JChar._JChar))
+  tripping c toJChar (fmap fromJChar)
+  where
+    fromJChar :: JChar HeXDigit -> Char
+    fromJChar = (JChar._JChar #)
+
+    toJChar :: Char -> Maybe (JChar HeXDigit)
+    toJChar = (^? JChar._JChar)
 
 prop_jnumber_scientific_prism :: Property
 prop_jnumber_scientific_prism = property $ do
   sci <- forAll $ Sci.scientific
     <$> Gen.integral (Range.linear 0 maxI)
     <*> Gen.int Range.linearBounded
-  Just sci === ((JNumber._JNumberScientific # sci) ^? JNumber._JNumberScientific)
+
+  tripping sci (JNumber._JNumberScientific #) (^? JNumber._JNumberScientific)
   where
     maxI :: Integer
     maxI = 2 ^ (32 :: Integer)
@@ -107,7 +119,7 @@ prop_tripping_int_list = property $ do
   xs <- forAll . Gen.list (Range.linear 0 100) $ Gen.int (Range.linear 0 9999)
   tripping xs
     (E.runPureEncoder (E.traversable E.int))
-    (D.simpleDecode Common.parseBS (D.list D.int) . BSL8.toStrict)
+    (D.simpleDecode (D.list D.int) Common.parseBS . BSL8.toStrict)
 
 prop_tripping_image_record_generic :: Property
 prop_tripping_image_record_generic = withTests 1 . property $
@@ -120,10 +132,6 @@ prop_tripping_newtype_fudge_generic = withTests 1 . property $
 prop_tripping_maybe_bool_generic :: Property
 prop_tripping_maybe_bool_generic = property $
   forAll (Gen.maybe Gen.bool) >>= Common.prop_generic_tripping
-
-prop_tripping_maybe_maybe_bool_generic :: Property
-prop_tripping_maybe_maybe_bool_generic = property $
-  forAll (Gen.maybe (Gen.maybe Gen.bool)) >>= Common.prop_generic_tripping
 
 prop_tripping_int_list_generic :: Property
 prop_tripping_int_list_generic = property $ do
@@ -154,7 +162,7 @@ prop_maybe_maybe = withTests 1 . property $ do
   where
     trippin' a = tripping a
       (E.runPureEncoder enc)
-      (D.simpleDecode Common.parseBS dec . BSL8.toStrict)
+      (D.simpleDecode dec Common.parseBS . BSL8.toStrict)
 
     enc = E.maybeOrNull' . E.mapLikeObj' . E.atKey' "boop"
       $ E.maybeOrNull' (E.mapLikeObj' (E.atKey' "beep" E.bool'))
