+++ waargonaut-deps.nix
@@ -20,7 +20,6 @@ in
 
   generics-sop   = self.callHackage "generics-sop" "0.3.2.0" {};
 
-  # basement       = self.callHackage "basement" "0.0.8" {};
   generic-lens = pkgs.haskell.lib.dontCheck super.generic-lens;
 
   hw-parser      = self.callHackage "hw-parser" "0.1.0.0" {};
